requirements.txt
# Project dependencies

torch>=2.0.0
torchvision>=0.15.0
kornia>=0.7.0
Pillow>=10.1.0
numpy>=1.21.0
matplotlib>=3.5.0
